requirements.txt
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.7.2
autopep8==2.0.1
bcrypt==4.0.1
cachetools==4.2.2
certifi==2022.12.7
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
Django==4.2.2
django-formtools==2.4.1
django-otp==1.2.2
django-phonenumber-field==6.4.0
django-two-factor-auth==1.15.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
fonttools==4.39.4
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
imageio==2.31.0
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
liac-arff==2.5.0
matplotlib==3.7.1
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
phonenumbers==8.13.15
Pillow==9.5.0
pipenv==2023.6.18
platformdirs==3.8.0
ply==3.11
pyaes==1.6.1
pyasn1==0.5.0
pycodestyle==2.10.0
pycryptodome==3.18.0
pygame==2.4.0
PyJWT==1.7.1
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.7
pyotp==2.8.0
pyparsing==3.0.9
pypng==0.20220715.0
python-dateutil==2.8.2
python-mnist==0.7
python-telegram-bot==13.9
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
qrcode==7.4.2
regex==2023.3.23
rfc3986==1.5.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Telethon==1.28.5
threadpoolctl==3.1.0
tifffile==2023.4.12
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
utils==1.0.1
vboxapi==1.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
whitenoise==6.5.0
